\name{gsDesignGUI}
\alias{gsDesignGUI}
\title{Graphical user interface functionality for gsDesign}
\description{
The primary function to launch the gsDesign GUI
}
\usage{gsDesignGUI(x)}

\arguments{
\item{x}{A named list of R objects representing a converted Qt design QMap.}  
}

\details{
The gsDesign GUI ...
}

\examples{
gsDesignGUI(gsDesignGUI:::fauxDesignList)
}

\keyword{interface}
